Refactor SCV component and introduce SCVBase and InventoryBase components
- Removed the SCV component's internal logic and replaced it with SCVBase for better separation of concerns.
- Created SCVBase component to handle the display of change history with support for user roles and actions.
- Introduced InventoryBase component to manage the inventory of assets, including filtering and loading logic.
- Added user service function to fetch current user information.
- Updated shared components index for easier imports.
- Enhanced error handling and loading states across components.
