+++ frontend/src/app/auditor/inventory/Inventory.jsx
@@ -1,166 +1,14 @@
-import React, { useState, useCallback } from "react";
-import { SearchBar, CardActivo } from "../../../components/ui";
+import React from "react";
+import { InventoryBase } from "../../../components/shared";
 
 const Inventory = ({ className = "", onNavigateToSCV, ...props }) => {
-    const [filteredActivos, setFilteredActivos] = useState([]);
-    const [isFiltered, setIsFiltered] = useState(false);
-
-    const handleFilter = useCallback((filters) => {
-        console.log('Filtros aplicados:', filters);
-        
-        // Si no hay filtros, mostrar todos los activos
-        if (!filters.name && !filters.category && !filters.estado) {
-            setFilteredActivos([]);
-            setIsFiltered(false);
-            return;
-        }
-
-        // Aplicar filtros
-        const filtered = activos.filter(activo => {
-            // Filtro por nombre, código o responsable
-            if (filters.name) {
-                const searchTerm = filters.name.toLowerCase();
-                const matchesName = activo.nombre.toLowerCase().includes(searchTerm);
-                const matchesCode = activo.codigo.toLowerCase().includes(searchTerm);
-                const matchesResponsible = activo.responsable.toLowerCase().includes(searchTerm);
-                
-                if (!matchesName && !matchesCode && !matchesResponsible) {
-                    return false;
-                }
-            }
-
-            // Filtro por categoría
-            if (filters.category && activo.categoria !== filters.category) {
-                return false;
-            }
-
-            // Filtro por estado
-            if (filters.estado && activo.estado !== filters.estado) {
-                return false;
-            }
-
-            return true;
-        });
-
-        setFilteredActivos(filtered);
-        setIsFiltered(true);
-    }, []);
-
-    const handleHistorialClick = (activo) => {
-        console.log('Navegando al historial de:', activo.nombre);
-        if (onNavigateToSCV) {
-            onNavigateToSCV(activo);
-        }
-    };
-
-    // Datos de ejemplo
-    const activos = [
-        {
-            nombre: "Servidor Web Principal",
-            codigo: "SWP-001",
-            estado: "Activo",
-            categoria: "Infraestructura",
-            descripcion: "Servidor web para aplicaciones corporativas con balanceador de carga",
-            responsable: "Abigail Flores",
-            version: "v1.0.0",
-            ubicacion: "Sala 5 - Dep TI",
-            fecha_creacion: "2025-11-23",
-            acciones_disponibles: ["Historial de Versiones"]
-        },
-        {
-            nombre: "Base de Datos MySQL",
-            codigo: "BDM-002",
-            estado: "Mantenimiento",
-            categoria: "Base de Datos",
-            descripcion: "Base de datos principal para almacenamiento de información corporativa",
-            responsable: "Javier Orellana",
-            version: "v2.1.3",
-            ubicacion: "Sala 3 - Dep TI",
-            fecha_creacion: "2025-11-20",
-            acciones_disponibles: ["Historial de Versiones"]
-        },
-        {
-            nombre: "Sistema de Backup",
-            codigo: "SBK-003",
-            estado: "Inactivo",
-            categoria: "Respaldo",
-            descripcion: "Sistema automatizado de respaldos nocturnos",
-            responsable: "Valeria Enriquez",
-            version: "v1.5.2",
-            ubicacion: "Sala 7 - Dep TI",
-            fecha_creacion: "2025-11-18",
-            acciones_disponibles: ["Historial de Versiones"]
-        },
-        {
-            nombre: "Firewall Corporativo",
-            codigo: "FWC-004",
-            estado: "En Revision",
-            categoria: "Seguridad",
-            descripcion: "Firewall de nueva generación para protección perimetral",
-            responsable: "Andrés Aguilar",
-            version: "v3.0.1",
-            ubicacion: "Sala 1 - Seguridad",
-            fecha_creacion: "2025-11-25",
-            acciones_disponibles: ["Historial de Versiones"]
-        }
-    ];
-
-    // Usar activos filtrados o todos los activos
-    const activosToShow = isFiltered ? filteredActivos : activos;
-
-    // Definir los campos de búsqueda
-    const searchFields = [
-        {
-            name: 'name',
-            label: 'Nombre',
-            type: 'text',
-            placeholder: 'Busque activo por nombre, código o responsable'
-        },
-        {
-            name: 'category',
-            label: 'Categoría',
-            type: 'select',
-            options: [
-                { value: 'Infraestructura', label: 'Infraestructura' },
-                { value: 'Base de Datos', label: 'Base de Datos' },
-                { value: 'Respaldo', label: 'Respaldo' },
-                { value: 'Seguridad', label: 'Seguridad' }
-            ]
-        },
-        {
-            name: 'estado',
-            label: 'Estado',
-            type: 'select',
-            options: [
-                { value: 'Activo', label: 'Activo' },
-                { value: 'Mantenimiento', label: 'Mantenimiento' },
-                { value: 'Inactivo', label: 'Inactivo' },
-                { value: 'En Revision', label: 'En Revisión' }
-            ]
-        }
-    ];
-
     return (
-        <div className={`inventory-page ${className}`} {...props}>
-            <div className="user-header">
-                <div className="user-header-text">
-                    <h2>Inventario de Activos</h2>
-                    <h6>{activosToShow.length} activos en total</h6>
-                </div>
-            </div>
-            
-            <SearchBar fields={searchFields} onFilter={handleFilter} />
-            
-            <div className="activos-grid">
-                {activosToShow.map((activo, index) => (
-                    <CardActivo
-                        key={index}
-                        activo={activo}
-                        onHistorialClick={handleHistorialClick}
-                    />
-                ))}
-            </div>
-        </div>
+        <InventoryBase
+            role="auditor"
+            className={className}
+            onNavigateToSCV={onNavigateToSCV}
+            {...props}
+        />
     );
 };
