+++ frontend/src/components/shared/index.js
@@ -1,0 +1,2 @@
+export { default as SCVBase } from './SCVBase';
+export { default as InventoryBase } from './InventoryBase';
